move cbor decoder into its own package
