+++ sdk/confidentialledger/azure-codetransparency/azure/codetransparency/cbor/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+# --------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------
+
+from ._decoder import CBORDecoder
+
+__all__ = ["CBORDecoder"]
