+++ sdk/confidentialledger/azure-codetransparency/tests/test_recorded_codetransparency_client.py
@@ -19,6 +19,8 @@ from devtools_testutils import (
 
 from azure.codetransparency import (
     CodeTransparencyClient,
+)
+from azure.codetransparency.cbor import (
     CBORDecoder,
 )
